+++ src/app/data-extracts/page.tsx
@@ -1,4 +1,8 @@
-"use client";
+import type { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "Data Extracts",
+};
 
 export default function DataExtractsPage() {
   return (
