+++ src/app/display-energy-certificate-and-advisory-report/certificate/[rrn]/page.tsx
@@ -13,6 +13,11 @@ import ArCertificate from "@/app/components/certificate/decar/ArCertificate";
 import DownloadButton from "@/app/components/DownloadButton";
 
 import { formatIsoDateLong, isExpiredDate } from "@/app/utils/date";
+import { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "Display Energy Certificate or Advisory Report Certificate",
+};
 
 type SummaryResponse = ApiEnvelope<DecarSummary>;
