+++ src/app/layout.tsx
@@ -1,9 +1,13 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 // import SiteHeader from "@scottish-government/designsystem-react/dist/components/SiteHeader/SiteHeader";
+
 import AppHeader from "./components/AppHeader";
 import AppFooter from "./components/AppFooter";
 // import "@scottish-government/design-system/dist/css/design-system.css";
+
+import AppBreadcrumbs from "./components/AppBreadcrumbs";
+import AppPhaseBanner from "./components/AppPhaseBanner";
 
 import "./globals.css";
 import "./globals-print.css";
@@ -21,7 +25,10 @@ const geistMono = Geist_Mono({
 });
 
 export const metadata: Metadata = {
-  title: "Energy Certificates",
+  title: {
+    default: "Energy Certificates",
+    template: "%s - Energy Certificates",
+  },
   description: "Find Scottish Energy Certificates",
 };
 
@@ -41,6 +48,8 @@ export default function RootLayout({
           <AppHeader />
 
           <main id="main-content" className="ds_page__middle">
+            <AppPhaseBanner />
+            <AppBreadcrumbs />
             {children}
           </main>
